move seen_region logic to a separate seen.[hc] module for testing. currently failing test:
  1) Failure (tests.report.test_lighthouse):
  ./scripts/tests/common.lua:1051: true expected but was false
