+++ src/bind_unit.c
@@ -18,6 +18,7 @@ without prior permission by the authors of Eressea.
 #include "bindings.h"
 #include "move.h"
 #include "reports.h"
+#include "seen.h"
 
 /*  attributes includes */
 #include <attributes/racename.h>
