+++ src/bindings.c
@@ -25,6 +25,7 @@ without prior permission by the authors of Eressea.
 #include "helpers.h"
 #include "console.h"
 #include "reports.h"
+#include "seen.h"
 
 #include <kernel/config.h>
