+++ src/jsreport.c
@@ -1,5 +1,6 @@
 ﻿#include "reports.h"
 #include "jsreport.h"
+#include "seen.h"
 #include <kernel/faction.h>
 #include <kernel/region.h>
 #include <kernel/terrain.h>
