+++ src/report.c
@@ -22,6 +22,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <kernel/config.h>
 
 #include "reports.h"
+#include "seen.h"
 #include "laws.h"
 #include "travelthru.h"
 #include "monster.h"
