+++ src/reports.c
@@ -20,6 +20,7 @@ OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
 #include <kernel/config.h>
 #include "reports.h"
 #include "laws.h"
+#include "seen.h"
 #include "travelthru.h"
 #include "lighthouse.h"
 
@@ -1090,121 +1091,6 @@ static void get_addresses(report_context * ctx)
     ctx->addresses = flist;
 }
 
-#define MAXSEEHASH 0x1000
-seen_region *reuse;
-
-seen_region **seen_init(void)
-{
-    return (seen_region **)calloc(MAXSEEHASH, sizeof(seen_region *));
-}
-
-void seen_done(seen_region * seehash[])
-{
-    int i;
-    for (i = 0; i != MAXSEEHASH; ++i) {
-        seen_region *sd = seehash[i];
-        if (sd == NULL)
-            continue;
-        while (sd->nextHash != NULL)
-            sd = sd->nextHash;
-        sd->nextHash = reuse;
-        reuse = seehash[i];
-        seehash[i] = NULL;
-    }
-    /*  free(seehash); */
-}
-
-void free_seen(void)
-{
-    while (reuse) {
-        seen_region *r = reuse;
-        reuse = reuse->nextHash;
-        free(r);
-    }
-}
-
-void
-link_seen(seen_region * seehash[], const region * first, const region * last)
-{
-    const region *r = first;
-    seen_region *sr = NULL;
-
-    if (first == last)
-        return;
-
-    do {
-        sr = find_seen(seehash, r);
-        r = r->next;
-    } while (sr == NULL && r != last);
-
-    while (r != last) {
-        seen_region *sn = find_seen(seehash, r);
-        if (sn != NULL) {
-            sr->next = sn;
-            sr = sn;
-        }
-        r = r->next;
-    }
-    if (sr) sr->next = 0;
-}
-
-seen_region *find_seen(struct seen_region *seehash[], const region * r)
-{
-    unsigned int index = reg_hashkey(r) & (MAXSEEHASH - 1);
-    seen_region *find = seehash[index];
-    while (find) {
-        if (find->r == r)
-            return find;
-        find = find->nextHash;
-    }
-    return NULL;
-}
-
-static void get_seen_interval(report_context * ctx)
-{
-    /* this is required to find the neighbour regions of the ones we are in,
-     * which may well be outside of [firstregion, lastregion) */
-    int i;
-
-    assert(ctx->f->seen);
-    for (i = 0; i != MAXSEEHASH; ++i) {
-        seen_region *sr = ctx->f->seen[i];
-        while (sr != NULL) {
-            if (ctx->first == NULL || sr->r->index < ctx->first->index) {
-                ctx->first = sr->r;
-            }
-            if (ctx->last != NULL && sr->r->index >= ctx->last->index) {
-                ctx->last = sr->r->next;
-            }
-            sr = sr->nextHash;
-        }
-    }
-    link_seen(ctx->f->seen, ctx->first, ctx->last);
-}
-
-bool
-add_seen(struct seen_region *seehash[], struct region *r, unsigned char mode,
-bool dis)
-{
-    seen_region *find = find_seen(seehash, r);
-    if (find == NULL) {
-        unsigned int index = reg_hashkey(r) & (MAXSEEHASH - 1);
-        if (!reuse)
-            reuse = (seen_region *)calloc(1, sizeof(struct seen_region));
-        find = reuse;
-        reuse = reuse->nextHash;
-        find->nextHash = seehash[index];
-        seehash[index] = find;
-        find->r = r;
-    }
-    else if (find->mode >= mode) {
-        return false;
-    }
-    find->mode = mode;
-    find->disbelieves |= dis;
-    return true;
-}
-
 typedef struct report_type {
     struct report_type *next;
     report_fun write;
@@ -1653,7 +1539,8 @@ static void prepare_report(struct report_context *ctx, faction *f)
             view(f->seen, r, f);
         }
     }
-    get_seen_interval(ctx);
+    get_seen_interval(ctx->f->seen, &ctx->first, &ctx->last);
+    link_seen(ctx->f->seen, ctx->first, ctx->last);
 }
 
 int write_reports(faction * f, time_t ltime)
